requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.3
Django==3.2.12
django-allauth==0.43.0
django-comments-xtd==2.9.5
django-contrib-comments==2.1.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
django-phrase==1.0.8
django-rest-framework==0.1.0
django-summernote==0.8.20.0
djangorestframework==3.13.1
docutils==0.18.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysqlclient==2.1.0
oauth2-provider==0.0
oauthlib==3.2.0
packaging==21.3
Pillow==7.2.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.7
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
whitenoise==6.0.0
